+++ packages/snippets/src/browser/getElementContentSize.js
@@ -1,8 +1,17 @@
+const setElementStyleProperties = require('./setElementStyleProperties')
+
 function getElementEntireSize([element] = []) {
-  return {
-    width: Math.max(element.clientWidth, element.scrollWidth),
-    height: Math.max(element.clientHeight, element.scrollHeight),
+  let originalStyleProperties
+  if (element === document.documentElement) {
+    originalStyleProperties = setElementStyleProperties([element, {transform: 'none'}])
   }
+
+  const size = {width: element.scrollWidth, height: element.scrollHeight}
+
+  if (originalStyleProperties) {
+    setElementStyleProperties([element, originalStyleProperties])
+  }
+  return size
 }
 
 module.exports = getElementEntireSize