+++ packages/screenshoter/CHANGELOG.md
@@ -3,8 +3,11 @@
 
 ## Unreleased
 
+- `crop`, `scale` and `rotate` now should be placed in `stabilization` object
+- rename `context` to `frames`
+- rename `isFully` to `fully`
+- integrate dom-capture
 - chore: add husky
-- add `Galaxy S20` device name to enum
 - fix bug when screenshots on iPad were taken with Safari navigation bar and iOS status bar
 
 ## 2.1.1 - 2021/1/15